# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Paper Setting API",
    "description": "Local-first DOCX thesis formatting compiler",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/templates/inspect": {
      "post": {
        "tags": [
          "templates"
        ],
        "summary": "Inspect Template Document",
        "operationId": "inspect_template_document_api_v1_templates_inspect_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_inspect_template_document_api_v1_templates_inspect_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateInspection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/templates/fill/preview": {
      "post": {
        "tags": [
          "templates"
        ],
        "summary": "Preview Template Fill",
        "operationId": "preview_template_fill_api_v1_templates_fill_preview_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_preview_template_fill_api_v1_templates_fill_preview_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateFillPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/templates/fill": {
      "post": {
        "tags": [
          "templates"
        ],
        "summary": "Generate Filled Template",
        "operationId": "generate_filled_template_api_v1_templates_fill_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_generate_filled_template_api_v1_templates_fill_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Filled DOCX copy",
            "content": {
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/templates/structure/preview": {
      "post": {
        "tags": [
          "templates"
        ],
        "summary": "Preview Template Structure",
        "operationId": "preview_template_structure_api_v1_templates_structure_preview_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_preview_template_structure_api_v1_templates_structure_preview_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateStructurePreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/templates/structure": {
      "post": {
        "tags": [
          "templates"
        ],
        "summary": "Generate Template Structure",
        "operationId": "generate_template_structure_api_v1_templates_structure_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_generate_template_structure_api_v1_templates_structure_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "DOCX copy with updated section structure",
            "content": {
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/templates/combined/preview": {
      "post": {
        "tags": [
          "templates"
        ],
        "summary": "Preview Template Combined",
        "operationId": "preview_template_combined_api_v1_templates_combined_preview_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_preview_template_combined_api_v1_templates_combined_preview_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateCombinedPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/templates/combined": {
      "post": {
        "tags": [
          "templates"
        ],
        "summary": "Generate Template Combined",
        "operationId": "generate_template_combined_api_v1_templates_combined_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_generate_template_combined_api_v1_templates_combined_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "DOCX copy with filled fields and updated section structure",
            "content": {
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_generate_filled_template_api_v1_templates_fill_post": {
        "properties": {
          "document": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Document",
            "description": "DOCX template to fill"
          },
          "values": {
            "type": "string",
            "title": "Values",
            "description": "JSON object keyed by placeholder or control tag"
          },
          "source_sha256": {
            "type": "string",
            "title": "Source Sha256"
          },
          "plan_version": {
            "type": "string",
            "title": "Plan Version"
          }
        },
        "type": "object",
        "required": [
          "document",
          "values",
          "source_sha256",
          "plan_version"
        ],
        "title": "Body_generate_filled_template_api_v1_templates_fill_post"
      },
      "Body_generate_template_combined_api_v1_templates_combined_post": {
        "properties": {
          "document": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Document",
            "description": "DOCX template to fill and restructure"
          },
          "values": {
            "type": "string",
            "title": "Values",
            "description": "JSON object keyed by placeholder or control tag"
          },
          "configuration": {
            "type": "string",
            "title": "Configuration",
            "description": "Section structure configuration JSON"
          },
          "source_sha256": {
            "type": "string",
            "title": "Source Sha256"
          },
          "plan_version": {
            "type": "string",
            "title": "Plan Version"
          }
        },
        "type": "object",
        "required": [
          "document",
          "values",
          "configuration",
          "source_sha256",
          "plan_version"
        ],
        "title": "Body_generate_template_combined_api_v1_templates_combined_post"
      },
      "Body_generate_template_structure_api_v1_templates_structure_post": {
        "properties": {
          "document": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Document",
            "description": "DOCX template to restructure"
          },
          "configuration": {
            "type": "string",
            "title": "Configuration",
            "description": "Section structure configuration JSON"
          },
          "source_sha256": {
            "type": "string",
            "title": "Source Sha256"
          },
          "plan_version": {
            "type": "string",
            "title": "Plan Version"
          }
        },
        "type": "object",
        "required": [
          "document",
          "configuration",
          "source_sha256",
          "plan_version"
        ],
        "title": "Body_generate_template_structure_api_v1_templates_structure_post"
      },
...
      "Body_inspect_template_document_api_v1_templates_inspect_post": {
        "properties": {
          "document": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Document",
            "description": "School or journal DOCX template"
          }
        },
        "type": "object",
        "required": [
          "document"
        ],
        "title": "Body_inspect_template_document_api_v1_templates_inspect_post"
      },
      "Body_preview_template_combined_api_v1_templates_combined_preview_post": {
        "properties": {
          "document": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Document",
            "description": "DOCX template to fill and restructure"
          },
          "values": {
            "type": "string",
            "title": "Values",
            "description": "JSON object keyed by placeholder or control tag"
          },
          "configuration": {
            "type": "string",
            "title": "Configuration",
            "description": "Section structure configuration JSON"
          },
          "source_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Sha256"
          }
        },
        "type": "object",
        "required": [
          "document",
          "values",
          "configuration"
        ],
        "title": "Body_preview_template_combined_api_v1_templates_combined_preview_post"
      },
      "Body_preview_template_fill_api_v1_templates_fill_preview_post": {
        "properties": {
          "document": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Document",
            "description": "DOCX template to fill"
          },
          "values": {
            "type": "string",
            "title": "Values",
            "description": "JSON object keyed by placeholder or control tag"
          },
          "source_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Sha256"
          }
        },
        "type": "object",
        "required": [
          "document",
          "values"
        ],
        "title": "Body_preview_template_fill_api_v1_templates_fill_preview_post"
      },
      "Body_preview_template_structure_api_v1_templates_structure_preview_post": {
        "properties": {
          "document": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Document",
            "description": "DOCX template to restructure"
          },
          "configuration": {
            "type": "string",
            "title": "Configuration",
            "description": "Section structure configuration JSON"
          },
          "source_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Sha256"
          }
        },
        "type": "object",
        "required": [
          "document",
          "configuration"
        ],
        "title": "Body_preview_template_structure_api_v1_templates_structure_preview_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TemplateCombinedPreview": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "plan_kind": {
            "type": "string",
            "const": "template_combined",
            "title": "Plan Kind",
            "default": "template_combined"
          },
          "source_filename": {
            "type": "string",
            "title": "Source Filename"
          },
          "source_sha256": {
            "type": "string",
            "title": "Source Sha256"
          },
          "plan_version": {
            "type": "string",
            "title": "Plan Version"
          },
          "fill": {
            "$ref": "#/components/schemas/TemplateFillPreview"
          },
          "structure": {
            "$ref": "#/components/schemas/TemplateStructurePreview"
          },
          "can_generate": {
            "type": "boolean",
            "title": "Can Generate"
          },
          "blockers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blockers"
          },
          "output_filename": {
            "type": "string",
            "title": "Output Filename"
          }
        },
        "type": "object",
        "required": [
          "source_filename",
          "source_sha256",
          "plan_version",
          "fill",
          "structure",
          "can_generate",
          "output_filename"
        ],
        "title": "TemplateCombinedPreview"
      },
...
      "TemplateFillPreview": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "source_filename": {
            "type": "string",
            "title": "Source Filename"
          },
          "source_sha256": {
            "type": "string",
            "title": "Source Sha256"
          },
          "plan_version": {
            "type": "string",
            "title": "Plan Version"
          },
          "operations": {
            "items": {
              "$ref": "#/components/schemas/TemplateFillOperation"
            },
            "type": "array",
            "title": "Operations"
          },
          "missing_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Keys"
          },
          "unused_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unused Keys"
          },
          "blocked_targets": {
            "items": {
              "$ref": "#/components/schemas/TemplateFillBlockedTarget"
            },
            "type": "array",
            "title": "Blocked Targets"
          },
          "replacement_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Replacement Count"
          },
          "can_generate": {
            "type": "boolean",
            "title": "Can Generate"
          },
          "output_filename": {
            "type": "string",
            "title": "Output Filename"
          }
        },
        "type": "object",
        "required": [
          "source_filename",
          "source_sha256",
          "plan_version",
          "operations",
          "replacement_count",
          "can_generate",
          "output_filename"
        ],
        "title": "TemplateFillPreview"
      },
...
      "TemplateInspection": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "source_filename": {
            "type": "string",
            "title": "Source Filename"
          },
          "source_sha256": {
            "type": "string",
            "title": "Source Sha256"
          },
          "summary": {
            "$ref": "#/components/schemas/TemplateSummary"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/TemplateSectionSummary"
            },
            "type": "array",
            "title": "Sections"
          },
          "styles": {
            "items": {
              "$ref": "#/components/schemas/TemplateStyleSummary"
            },
            "type": "array",
            "title": "Styles"
          },
          "placeholders": {
            "items": {
              "$ref": "#/components/schemas/PlaceholderCandidate"
            },
            "type": "array",
            "title": "Placeholders"
          },
          "content_controls": {
            "items": {
              "$ref": "#/components/schemas/ContentControlCandidate"
            },
            "type": "array",
            "title": "Content Controls"
          },
          "semantic_candidates": {
            "items": {
              "$ref": "#/components/schemas/SemanticMappingCandidate"
            },
            "type": "array",
            "title": "Semantic Candidates"
          },
          "field_instructions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Field Instructions"
          },
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/TemplateCapabilityFinding"
            },
            "type": "array",
            "title": "Capabilities"
          },
          "even_and_odd_headers": {
            "type": "boolean",
            "title": "Even And Odd Headers",
            "default": false
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "source_filename",
          "source_sha256",
          "summary",
          "sections",
          "styles",
          "placeholders",
          "content_controls",
          "semantic_candidates",
          "field_instructions",
          "capabilities"
        ],
        "title": "TemplateInspection"
      },
...
      "TemplateStructurePreview": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "source_filename": {
            "type": "string",
            "title": "Source Filename"
          },
          "source_sha256": {
            "type": "string",
            "title": "Source Sha256"
          },
          "plan_version": {
            "type": "string",
            "title": "Plan Version"
          },
          "configuration": {
            "$ref": "#/components/schemas/TemplateSectionStructureConfig"
          },
          "operations": {
            "items": {
              "$ref": "#/components/schemas/TemplateStructureOperation"
            },
            "type": "array",
            "title": "Operations"
          },
          "blockers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blockers"
          },
          "can_generate": {
            "type": "boolean",
            "title": "Can Generate"
          },
          "output_filename": {
            "type": "string",
            "title": "Output Filename"
          }
        },
        "type": "object",
        "required": [
          "source_filename",
          "source_sha256",
          "plan_version",
          "configuration",
          "operations",
          "can_generate",
          "output_filename"
        ],
        "title": "TemplateStructurePreview"
      },
...
      "TemplateStyleSummary": {
        "properties": {
          "style_id": {
            "type": "string",
            "title": "Style Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "style_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Style Type"
          },
          "based_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Based On"
          },
          "usage_count": {
            "type": "integer",
            "title": "Usage Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "style_id"
        ],
        "title": "TemplateStyleSummary"
      },
...
    }
  }
}